requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.13
django-admin-honeypot-updated-2021==1.2.0
django-cleanup==8.1.0
django-database-url==1.0.3
django-formtools==2.5.1
django-ipware==7.0.1
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-two-factor-auth==1.16.0
django-widget-tweaks==1.5.0
gunicorn==22.0.0
idna==3.7
packaging==24.0
phonenumbers==8.13.35
pillow==10.3.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-decouple==3.8
python-ipware==3.0.0
qrcode==7.4.2
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
